% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{DT_runif}
\alias{DT_runif}
\title{data.table runif}
\usage{
DT_runif(n, name, lower, upper)
}
\arguments{
\item{n}{number of observations}

\item{name}{variable names}

\item{lower}{lower bounds}

\item{upper}{upper bounds}
}
\value{
a data.table of hyperparameters
}
\description{
Generates random variables for multiple hyperparameters in data.table format.
}
\keyword{internal}
