% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{DT_Top_K}
\alias{DT_Top_K}
\title{data.table Top_K}
\usage{
DT_Top_K(DT, Value_Var, K, maximize)
}
\arguments{
\item{DT}{data.table}

\item{Value_Var}{variable for selecting Top K row(s)}

\item{K}{number of K}

\item{maximize}{maximizing indicator}
}
\value{
a data.table of hyperparameters
}
\description{
Select Top K rows from data.table based on certain variable.
}
\keyword{internal}
