% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hyperband_Exp.R
\name{Hyperband_Exp}
\alias{Hyperband_Exp}
\title{Hyperband Experiment}
\usage{
Hyperband_Exp(R, R_unit, eta = 3)
}
\arguments{
\item{R}{Resource parameter, the maximum amount of resource that can be allocated to a single hyperparameter configuration.}

\item{R_unit}{Resource unit, the minimum amount of computation where different hyperparameter configurations start to separate.}

\item{eta}{an input that controls the proportion of configurations discarded in each round of SuccessiveHalving.}
}
\description{
Hyperband algorithm experiment function.
}
\examples{
Hyperband_Exp(R = 81, R_unit = 1, eta = 3)
}
