% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rHyperband.R
\docType{package}
\name{rHyperband}
\alias{rHyperband}
\alias{rHyperband-package}
\title{rHyperband: Hyperband - A Novel Bandit-Based Approach to Hyperparameter Optimization}
\description{
A Pure R implementation of Hyperband algorithm for Hyperparameter Optimization.
}
